+++ .cirrus/setup.sh
@@ -10,7 +10,7 @@ BASE_FOLDER="${BASE_FOLDER:-"/root/.gcp/cache"}"
 
 mkdir -p ${BASE_FOLDER}
 
-curl -LO https://dl.google.com/dl/cloudsdk/channels/rapid/downloads/google-cloud-cli-${GCLOUD_CLI_VERSION}-linux-x86_64.tar.gz
+curl -LO --proto "=https" https://dl.google.com/dl/cloudsdk/channels/rapid/downloads/google-cloud-cli-${GCLOUD_CLI_VERSION}-linux-x86_64.tar.gz
 echo "${GCLOUD_CLI_CHECKSUM}  google-cloud-cli-${GCLOUD_CLI_VERSION}-linux-x86_64.tar.gz" | sha256sum -c
 tar -xf google-cloud-cli-${GCLOUD_CLI_VERSION}-linux-x86_64.tar.gz
 chmod +x ./google-cloud-sdk
@@ -19,7 +19,7 @@ rm -rf google-cloud-cli-${GCLOUD_CLI_VERSION}-linux-x86_64.tar.gz
 
 gcloud components install gke-gcloud-auth-plugin kubectl --quiet
 
-curl -LO https://github.com/GoogleCloudPlatform/marketplace-k8s-app-tools/archive/refs/tags/${MPDEV_VERSION}.tar.gz
+curl -LO --proto "=https" https://github.com/GoogleCloudPlatform/marketplace-k8s-app-tools/archive/refs/tags/${MPDEV_VERSION}.tar.gz
 echo "${MPDEV_CHECKSUM}  ${MPDEV_VERSION}.tar.gz" | sha256sum -c
 tar -xf ${MPDEV_VERSION}.tar.gz
 mv marketplace-k8s-app-tools-${MPDEV_VERSION}/scripts/dev ${BASE_FOLDER}/mpdev