+++ charts/sonarqube-dce/CHANGELOG.md
@@ -6,6 +6,7 @@ All changes to this chart will be documented in this file.
 * Update Chart's version to 2025.5.0
 * Update the image and readinessProbe used by postgresql after they migrated to a legacy repository
 * Support Kubernetes v1.34.0
+* Disable Postgresql when the JDBC overwrite is used
 
 ## [2025.4.0]
 * Update Chart's version to 2025.4.0